k8s: refactored some k8s code out of the Cilium utils pkg
